Changes needed for MAX communication
The RF22 hardware doesn't use the same dewhitening sequence as the MAX
devices (even though the RF22 documentation claims it does), so we need
to do the dewhitening in software (outside of the RF22 library).
However, this also means the hardware cannot read the packet length, so
we apply a workaround to also find out the packet length in software and
tell the hardware about it halfway during the packet reception.